backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
supabase==2.5.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.27.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
pytest==9.0.2
pytest-asyncio==1.3.0
